requirements.txt
-e git://github.com/duckietown/duckietown-shell.git@v2#egg=duckietown-shell
ruamel.ordereddict
ruamel.yaml
